-Upgraded api endpoints. -Added withdraw, deposit and balance check -Improved compatibility with balance

+++ backend/app/models/account.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Enum, Float, ForeignKey, Integer, String
+from sqlalchemy import Column, Enum, Numeric, ForeignKey, Integer, String
 from sqlalchemy.orm import relationship
 
 from app.models import Base
@@ -8,7 +8,7 @@ class Account(Base):
     __tablename__ = "accounts"
 
     id = Column(Integer, primary_key=True, index=True)
-    balance = Column(Float, nullable=False)
+    balance = Column(Numeric(12, 2), nullable=False)
     currency = Column(Enum("GBP", "EUR", "USD", "JPY", name="currency"), nullable=False)
     type = Column(Enum("savings", "checking", name="account_type"), nullable=False)
     user_id = Column(Integer, ForeignKey("users.id"))
